pyproject.toml
[tool.poetry]
name = "youtube-bot-backend"
version = "0.1.0"
description = ""
authors = ["skip"]
license = "MIT"
readme = "README.md"
packages = [{include = "youtube_bot_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
mysqlclient = "^2.2.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.1.0"
requests = "^2.31.0"
django-environ = "^0.10.0"
dj-database-url = "^2.0.0"
psycopg2-binary = "^2.9.6"
whitenoise = "^6.5.0"
gunicorn = "^20.1.0"
google-api-python-client = "^2.91.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
